refactor(web): 파일이름을 kebab case로 변경
JIP-57
