+++ apps/web/src/views/register/user-info/index.tsx
@@ -16,7 +16,7 @@ const BIRTH_FORMAT_LEN = 6
 const emailRegex =
   /^[_A-Za-z0-9-\\+]+(\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\.[A-Za-z0-9]+)*(\.[A-Za-z]{2,})$/
 
-export default function RegisterPage() {
+export default function UserInfoPage() {
   const {
     control,
     formState: { touchedFields },
