+++ apps/web/src/views/register/user-info/styled.tsx
